+++ src/main/java/in/koreatech/koin/domain/community/article/dto/ArticleResponse.java
@@ -1,18 +1,18 @@
-package in.koreatech.koin.domain.community.articles.dto;
+package in.koreatech.koin.domain.community.article.dto;
 
-import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.community.articles.model.Article;
+import in.koreatech.koin.domain.community.article.model.Article;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record HotArticleItemResponse(
+public record ArticleResponse(
 
     @Schema(description = "게시글 고유 ID", example = "1", requiredMode = REQUIRED)
     Integer id,
@@ -23,11 +23,14 @@ public record HotArticleItemResponse(
     @Schema(description = "제목", example = "제목", requiredMode = REQUIRED)
     String title,
 
+    @Schema(description = "내용", example = "내용", requiredMode = REQUIRED)
+    String content,
+
     @Schema(description = "작성자 닉네임", example = "닉네임", requiredMode = REQUIRED)
     String nickname,
 
     @Schema(description = "조회수", example = "1", requiredMode = REQUIRED)
-    Integer hit,
+    int hit,
 
     @Schema(description = "생성 일자", example = "2023-01-04 12:00:01", requiredMode = REQUIRED)
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime createdAt,
@@ -36,11 +39,12 @@ public record HotArticleItemResponse(
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime updatedAt
 ) {
 
-    public static HotArticleItemResponse from(Article article) {
-        return new HotArticleItemResponse(
+    public static ArticleResponse of(Article article) {
+        return new ArticleResponse(
             article.getId(),
             article.getBoard().getId(),
             article.getTitle(),
+            article.getContent(),
             article.getNickname(),
             article.getHit(),
             article.getCreatedAt(),
