+++ src/main/java/in/koreatech/koin/domain/community/article/dto/ArticlesResponse.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.community.articles.dto;
+package in.koreatech.koin.domain.community.article.dto;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.community.articles.model.Article;
+import in.koreatech.koin.domain.community.article.model.Article;
 import in.koreatech.koin.global.model.Criteria;
 import io.swagger.v3.oas.annotations.media.Schema;
