+++ src/main/java/in/koreatech/koin/domain/community/article/repository/ArticleSearchLogRepository.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.community.articles.repository;
+package in.koreatech.koin.domain.community.article.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.community.articles.model.ArticleSearchLog;
+import in.koreatech.koin.domain.community.article.model.ArticleSearchLog;
 
 public interface ArticleSearchLogRepository extends Repository<ArticleSearchLog, Integer> {
