+++ src/main/java/in/koreatech/koin/domain/community/article/repository/BoardRepository.java
@@ -1,11 +1,11 @@
-package in.koreatech.koin.domain.community.articles.repository;
+package in.koreatech.koin.domain.community.article.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.community.articles.exception.ArticleNotFoundException;
-import in.koreatech.koin.domain.community.articles.model.Board;
+import in.koreatech.koin.domain.community.article.exception.ArticleNotFoundException;
+import in.koreatech.koin.domain.community.article.model.Board;
 
 public interface BoardRepository extends Repository<Board, Integer> {
     Optional<Board> findById(Integer id);
